pyproject.toml
[tool.poetry]
name = "care"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.0"
networkx = "^3.3"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
pygraphviz = "^1.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
